requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
RPi.GPIO==0.7.1
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
importlib-metadata==6.8.0
zipp==3.17.0